ref(onboarding): Refactor onboarding components for rendering in drawer
