+++ static/app/components/onboardingWizard/useOnboardingTasks.tsx
@@ -1,3 +1,7 @@
+import {useContext} from 'react';
+
+import {OnboardingContext} from 'sentry/components/onboarding/onboardingContext';
+import {getMergedTasks} from 'sentry/components/onboardingWizard/taskConfig';
 import {findCompleteTasks, taskIsDone} from 'sentry/components/onboardingWizard/utils';
 import {
   type OnboardingTask,
@@ -6,6 +10,7 @@ import {
 } from 'sentry/types/onboarding';
 import {useApiQuery} from 'sentry/utils/queryClient';
 import useOrganization from 'sentry/utils/useOrganization';
+import useProjects from 'sentry/utils/useProjects';
 
 // Merge supported onboarding tasks with their completion status from the server.
 function mergeTasks({
@@ -31,13 +36,7 @@ function mergeTasks({
 
 // This hook polls the onboarding tasks endpoint every second until all supported tasks are complete and
 // returns the task groups: "allTasks", "gettingStartedTasks", "beyondBasicsTasks", and "completeTasks".
-export function useOnboardingTasks({
-  supportedTasks,
-  enabled,
-}: {
-  enabled: boolean;
-  supportedTasks: OnboardingTask[];
-}): {
+export function useOnboardingTasks({disabled = false}: {disabled?: boolean} = {}): {
   allTasks: OnboardingTask[];
   beyondBasicsTasks: OnboardingTask[];
   completeTasks: OnboardingTask[];
@@ -46,12 +45,22 @@ export function useOnboardingTasks({
   refetch: () => void;
 } {
   const organization = useOrganization();
+  const onboardingContext = useContext(OnboardingContext);
+  const {projects} = useProjects();
+  const supportedTasks = getMergedTasks({
+    organization,
+    projects,
+    onboardingContext,
+  }).filter(task => task.display);
+
+  const allTasksDone = supportedTasks.every(findCompleteTasks);
 
   const {data: serverTasks = {onboardingTasks: []}, refetch} = useApiQuery<{
     onboardingTasks: OnboardingTaskStatus[];
   }>([`/organizations/${organization.slug}/onboarding-tasks/`], {
     staleTime: 0,
-    enabled,
+    enabled:
+      !!organization.features?.includes('onboarding') && !allTasksDone && !disabled,
     refetchInterval: query => {
       const data = query.state.data?.[0]?.onboardingTasks;
       if (!data) {
