+++ static/app/components/sidebar/onboardingStatus.tsx
@@ -4,11 +4,8 @@ import {css, useTheme} from '@emotion/react';
 import styled from '@emotion/styled';
 
 import GuideAnchor from 'sentry/components/assistant/guideAnchor';
-import {OnboardingContext} from 'sentry/components/onboarding/onboardingContext';
-import {OnboardingSidebar} from 'sentry/components/onboardingWizard/sidebar';
-import {getMergedTasks} from 'sentry/components/onboardingWizard/taskConfig';
+import {LegacyOnboardingSidebar} from 'sentry/components/onboardingWizard/sidebar';
 import {useOnboardingTasks} from 'sentry/components/onboardingWizard/useOnboardingTasks';
-import {findCompleteTasks} from 'sentry/components/onboardingWizard/utils';
 import ProgressRing, {
   RingBackground,
   RingBar,
@@ -22,7 +19,6 @@ import {isDemoModeEnabled} from 'sentry/utils/demoMode';
 import {useLocalStorageState} from 'sentry/utils/useLocalStorageState';
 import useMutateUserOptions from 'sentry/utils/useMutateUserOptions';
 import useOrganization from 'sentry/utils/useOrganization';
-import useProjects from 'sentry/utils/useProjects';
 import {useUser} from 'sentry/utils/useUser';
 import {useOnboardingSidebar} from 'sentry/views/onboarding/useOnboardingSidebar';
 
@@ -43,8 +39,6 @@ export function OnboardingStatus({
   const {mutate: mutateUserOptions} = useMutateUserOptions();
   const {activateSidebar} = useOnboardingSidebar();
   const organization = useOrganization();
-  const onboardingContext = useContext(OnboardingContext);
-  const {projects} = useProjects();
   const {shouldAccordionFloat} = useContext(ExpandedContext);
   const [quickStartCompleted, setQuickStartCompleted] = useLocalStorageState(
     `quick-start:${organization.slug}:completed`,
@@ -54,12 +48,6 @@ export function OnboardingStatus({
   const isActive = currentPanel === SidebarPanelKey.ONBOARDING_WIZARD;
   const demoMode = isDemoModeEnabled();
 
-  const supportedTasks = getMergedTasks({
-    organization,
-    projects,
-    onboardingContext,
-  }).filter(task => task.display);
-
   const {
     allTasks,
     gettingStartedTasks,
@@ -68,11 +56,7 @@ export function OnboardingStatus({
     completeTasks,
     refetch,
   } = useOnboardingTasks({
-    supportedTasks,
-    enabled:
-      !!organization.features?.includes('onboarding') &&
-      !supportedTasks.every(findCompleteTasks) &&
-      isActive,
+    disabled: !isActive,
   });
 
   const label = demoMode ? t('Guided Tours') : t('Onboarding');
@@ -189,7 +173,7 @@ export function OnboardingStatus({
         )}
       </Container>
       {isActive && (
-        <OnboardingSidebar
+        <LegacyOnboardingSidebar
           orientation={orientation}
           collapsed={collapsed}
           onClose={hidePanel}
